Add files via upload
This concept was started by my 17 year old and came from discussions of potential interests and uses for this structure.  

They wanted a Social platform that allowed them to pick and choose how they participated, what services they used, and the ability to sell items on their own without having to pay someone to host a market for them that would also give them a URL to send people to purchase items and pay them how they want to be paid. 

oh.... And they don't want to be mined at least not easily. 

This is the initial commit for a concept of an encrypted extendable social media platform.